+++ packages/@expo/metro-config/src/serializer/serializeChunks.ts
@@ -588,8 +588,10 @@ function collectOutputReferences(modules: Iterable<Module>, key: string): string
       [...modules]
         .map((module) => {
           return module.output.map((output) => {
-            if (key in output.data && typeof output.data[key] === 'string') {
-              return output.data[key];
+            // TODO: This is a mess. This needs to be properly typed
+            const data = output.data as any;
+            if (key in data && typeof data[key] === 'string') {
+              return data[key];
             }
             return undefined;
           });
@@ -618,7 +620,7 @@ function gatherChunks(
   settings: ChunkSettings,
   preModules: readonly Module[],
   graph: ReadOnlyGraph,
-  options: SerializerOptions<MixedOutput>,
+  options: SerializerOptions,
   isAsync: boolean = false,
   isEntry: boolean = false
 ): Set<Chunk> {
@@ -716,7 +718,7 @@ function removeEntryDepsFromAsyncChunks(entryChunk: Chunk, chunks: Set<Chunk>): 
 function extractCommonChunk(
   chunks: Set<Chunk>,
   graph: ReadOnlyGraph,
-  options: SerializerOptions<MixedOutput>
+  options: SerializerOptions
 ): Chunk | undefined {
   const toCompare = [...chunks.values()];
 
@@ -777,7 +779,7 @@ function createRuntimeChunk(
   entryChunk: Chunk,
   chunks: Set<Chunk>,
   graph: ReadOnlyGraph,
-  options: SerializerOptions<MixedOutput>
+  options: SerializerOptions
 ): void {
   const runtimeChunk = new Chunk('/__expo-metro-runtime.js', [], graph, options, false, true);