+++ packages/expo-file-system/android/src/main/java/expo/modules/filesystem/FileSystemModule.kt
@@ -13,6 +13,7 @@ import expo.modules.kotlin.modules.ModuleDefinition
 import expo.modules.kotlin.services.FilePermissionService
 import expo.modules.kotlin.typedarray.TypedArray
 import expo.modules.kotlin.types.Either
+import kotlinx.coroutines.runBlocking
 import java.io.File
 import java.net.URI
 
@@ -193,12 +194,24 @@ class FileSystemModule : Module() {
         file.creationTime
       }
 
-      Function("copy") { file: FileSystemFile, destination: FileSystemPath, options: RelocationOptions? ->
+      AsyncFunction("copy") Coroutine { file: FileSystemFile, destination: FileSystemPath, options: RelocationOptions? ->
         file.copy(destination, options ?: RelocationOptions())
       }
 
-      Function("move") { file: FileSystemFile, destination: FileSystemPath, options: RelocationOptions? ->
+      Function("copySync") { file: FileSystemFile, destination: FileSystemPath, options: RelocationOptions? ->
+        runBlocking {
+          file.copy(destination, options ?: RelocationOptions())
+        }
+      }
+
+      AsyncFunction("move") Coroutine { file: FileSystemFile, destination: FileSystemPath, options: RelocationOptions? ->
         file.move(destination, options ?: RelocationOptions())
+      }
+
+      Function("moveSync") { file: FileSystemFile, destination: FileSystemPath, options: RelocationOptions? ->
+        runBlocking {
+          file.move(destination, options ?: RelocationOptions())
+        }
       }
 
       Function("rename") { file: FileSystemFile, newName: String ->
@@ -294,12 +307,24 @@ class FileSystemModule : Module() {
         directory.validatePath()
       }
 
-      Function("copy") { directory: FileSystemDirectory, destination: FileSystemPath, options: RelocationOptions? ->
+      AsyncFunction("copy") Coroutine { directory: FileSystemDirectory, destination: FileSystemPath, options: RelocationOptions? ->
         directory.copy(destination, options ?: RelocationOptions())
       }
 
-      Function("move") { directory: FileSystemDirectory, destination: FileSystemPath, options: RelocationOptions? ->
+      Function("copySync") { directory: FileSystemDirectory, destination: FileSystemPath, options: RelocationOptions? ->
+        runBlocking {
+          directory.copy(destination, options ?: RelocationOptions())
+        }
+      }
+
+      AsyncFunction("move") Coroutine { directory: FileSystemDirectory, destination: FileSystemPath, options: RelocationOptions? ->
         directory.move(destination, options ?: RelocationOptions())
+      }
+
+      Function("moveSync") { directory: FileSystemDirectory, destination: FileSystemPath, options: RelocationOptions? ->
+        runBlocking {
+          directory.move(destination, options ?: RelocationOptions())
+        }
       }
 
       Function("rename") { directory: FileSystemDirectory, newName: String ->